+++ modyn/common/trigger_sample/include/trigger_sample_storage_wrapper.hpp
@@ -3,17 +3,18 @@
 
 extern "C" {
 
-int get_num_samples_in_file(const char* filename);
+long get_num_samples_in_file(const char* filename);
 
-void* get_worker_samples(const char* folder, int* size, const char* pattern, int start_index, int worker_subset_size);
-void* get_all_samples(const char* folder, int* size, const char* pattern);
+void* get_worker_samples(const char* folder, long* size, const char* pattern, long start_index,
+                         long worker_subset_size);
+void* get_all_samples(const char* folder, long* size, const char* pattern);
 
-void* parse_file(const char* filename, int* size);
+void* parse_file(const char* filename, long* size);
 
-void write_file(const char* filename, const void* data, int data_offset, int data_length, const char* header,
-                int header_length);
-void write_files(const char* filenames[], const void* data, int data_lengths[], const char* headers[],
-                 int header_length, std::uint64_t num_files);
+void write_file(const char* filename, const void* data, long data_offset, long data_length, const char* header,
+                long header_length);
+void write_files(const char* filenames[], const void* data, long data_lengths[], const char* headers[],
+                 long header_length, std::uint64_t num_files);
 
 void release_data(void* data);
 }
